requirements.txt
streamlit==1.38.0
tensorflow==2.20.0
numpy
Pillow
